requirements.txt
google-generativeai
Pillow
pydantic
